% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Chaos01}
\alias{Chaos01}
\alias{.onUnload}
\alias{Chaos01-package}
\title{Chaos01}
\usage{
.onUnload(libpath)
}
\description{
Computes and visualize the results of the 0-1 test for chaos proposed
by Gottwald and Melbourne (2004) <DOI:10.1137/080718851>. The algorithm is
available in parallel for the independent values of parameter c. Additionally,
fast RQA is added to distinguish chaos from noise.
}
\author{
Tomas Martinovic
}
\keyword{internal}
