% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos01_plot.R
\name{plot.chaos01}
\alias{plot.chaos01}
\title{Plot the additional results of 0-1 test for chaos.}
\usage{
\method{plot}{chaos01}(x, plotvar = c("PQ", "MD"), mdcol = NULL, col = 1,
  ...)
}
\arguments{
\item{x}{the object of "chaos01" class, produced by test.chaos01 function when parameter Val = "TRUE". Subset the output of the function to get the results for the concrete c. See the example.}

\item{plotvar}{list/vector define what should be plotted.
\itemize{
    \item c("PQ", "MD") - both plots Pc - Qc, and Mc/Dc should be plotted.
    \item "PQ"          - only Pc-Qc should be plotted.
    \item "MD"          - only Mc/Dc plot should be plotted.
}
Default is c("PQ", "MD").}

\item{mdcol}{vector of length 2 or NULL. 
If NULL colors in MD plot will be the same as in 'col' argument. If vector of length 2, first color stands for the Mc line and second color for the Dc line.
\itemize{
    \item NULL   - use color defined in 'col' argument.
    \item c(4,3) - use blue for the Mc line and green for the Dc line.
    \item c("firebrick", "cadetblue") - use of color names is also possible.
    }
    
When used, it overrides 'col' argument. It is possible to set colors as numbers, or by the string name. 
 
Default is NULL.}

\item{col}{color for the lines in plots as defined in plot(). Default is 1.}

\item{...}{arguments to be passed as graphical parameters.}
}
\description{
This function plot the Pc to Qc plot and Mc/Dc plot as described in Gottwald and Melbourne (2004).
}
\examples{
vec.x <- gen.logistic(mu = 3.55, iter = 2000)

#Output for each value of c
res2 <- testChaos01(vec.x, out = TRUE)

plot(res2[[1]], plotvar = c("PQ", "MD"), mdcol = c(4,3))
}
\references{
Gottwald G.A. and Melbourne I. (2004) On the implementation of the 0-1 Test for Chaos, SIAM J. Appl. Dyn. Syst., 8(1), 129–145.
}
\seealso{
\code{\link{testChaos01}}, \code{\link{plot.chaos01.res}}
}
\keyword{chaos}
\keyword{plot}
\keyword{test}
